refactor: Modify name Apple to Coin
